Add queue support
Adds support for song queues based on guild id. It's unlikely this bot
will be used in another guild, but I figured i would add support for it.
Creates a GuildQueues object which is used to store queue data as well as
references to the underlying audio player and associated resources on a
per guild basis. Handles adding songs to queue, playing the next song,
and cleaning up resources. Note that we don't store the VoiceConnection
object because discord.js does this for  us.