fix bug causing bot to fail when skipping tracks
I'm still not 100% sure why my previous implementation failed. The skip
command would cause the bot to freeze and then disconnect without
calling the cleanup function or sending any errors. Voicebot would also start to consume an
insane amount of memory. I had a hunch it was the list changes and it
seems like i was correct. Calling abort on the list's abort controller
after the promise race completes seems to be the issue. The reason i was
calling abort was to ensure that the timer being used for brpop timeout
would cancel after the promise race. It was causing the jest tests to
hang. Now, I'm just using one event listener with an abort controller
and a clearable timeout callback.