Catch ytdl errors
We ran into and issue where voicebot would crash if ytdl() failed to
download. More specifically, ytdl was attempting to trim a non-existent
id string. This change catches that error, logs it, and passes it to the
caller. In our case, voicebot checks for errors from the
download function in playNext, so it will gracefully handle this new throw(err)